Add support for v2 voters
