+++ services/core/shared/core/compat/sdk_v5/blocks.js
@@ -19,7 +19,7 @@ const BluebirdPromise = require('bluebird');
 const coreApi = require('./coreApi');
 const config = require('../../../../config');
 const { indexAccountsbyPublicKey } = require('./accounts');
-const { indexVotes } = require('./votes');
+const { indexVotes } = require('./voters');
 const { getIndexedAccountByPublicKey } = require('./helper');
 const { indexTransactions } = require('./transactions');
 const { getApiClient, parseToJSONCompatObj } = require('../common');
